fix creation/deletion race in mqueue::Dispatcher
The class was obviously assuming that intrusive_ptr would be thread
safe, which we know is not true on deletion/creation races. It would
lead to crashes in cases where the last connection would be closed
at the same time than new connections would appear.

The fix I've decided to pick is to simply keep the dispatcher alive.
It is a single thread, unique per RTT process. I am essentially assuming
that if someone tries to create MQ connections, he/she will try again
and that deleting the dispatcher thread provides very little.

In my case, it started appearing on a system where the MQ was not well
configured (low number of MQueues available per process). In this
configuration, Syskit would attempt to create an MQ, it would fail and
then would fallback to CORBA. Since Syskit parallelizes connection
creation, the situation that would trigger this bug appeared often enough