transition to Exception only after the hooks have been run
exception() was setting both TargetState and State to Exception
on entering. This meant that an external observer had no chance
to synchronize on the actual end of transition (i.e. Exception
would be returned by getState() while the component could be
still running).

This could cause synchronization issues if e.g. the external observer
was trying to recover() and configure(). In this case, the RTT runtime
would happily let it recover and also configure, **concurrently to having
the stopHook/cleanupHook/exceptionHook called by exception()** since
configure() does not synchronize with the component's thread.