feat: add a virtual setTaskStates method to TaskCore to "hook" into state changes
orogen currently hooks into the state change methods (start(),
configure() ...) to publish state changes to the state port, but
that is rather limited.

Notably, there is a race in the update-on-trigger behavior that
still is part of RTT. States might be queued in the wrong order
if the scheduler decides to yield to the activity thread rather
than finish start(). For instance, in the following timeline,
the state port will report STOPPED, EXCEPTION, RUNNING

- start() calls startHook
- start() changes the taskcontext state to Running
- start() calls trigger()
- scheduler yield to the activity
- the activity calls updateHook()
- updateHook() changes state to e.g. Exception
- start() finishes and publishes Running