feat: allow disabling the signalling remote call in pull corba connections
When the connection is bad or simply breaks, the signalling call can block, even
though it is defined oneway - the TCP connection still needs to go through. This
actually blocks any other connection that is on the same corba dispatcher, with
catastrophic results.

Since orocos.rb/async and vizkit do not use the functionality, provide the
possibility to disable signalling altogether.