chore: make it explicit which of the channel building calls actually update the policy
Policy updating is needed to exfiltrate some information in the OOB transport
case (namely, a name that explains what the other side should do to connect, as
for instance the MQ name for the MQ transport). Turns out that only the output
half is doing so, and the other take the policy as input.

Ideally, we would also have cleaned up what information is or is not being
passed to the other calls (the connect calls, for instance, really don't need
much policy information), but that would be for another PR.