fix: restore the old behaviour of Instance/Release
After a discussion on GitHub, it turns out that the initial code was indeed
breaking possibly existing code built around Instance: Instance creates the
instance and Release deletes it, regardless of how many times Instance has been
called. The new refcounting implementation was breaking that assumption

Renamed the "new" Release into Deref, and re-created the old behaviour on top of
the new.