fix: do stop the dispatcher thread in Deref before deleting it
RTT is littered with code that attempts to do cleanup in destructors, of things
that *very probably* have references to a subclass of the code.

In this instance, the dispatcher thread might be within the dispatcher code
itself, which might be referring to fields from the dispatcher subclass (and not
Activity). Assuming that `~Activity` can do cleanup cleanly is idiotic.

We should actually put assertions in destructors that check that cleanup has
been made ... but that's another longer term story.