fix: ensure that calling exception() will not call stopHook or cleanupHook twice
To my surprise (!), we found out that calling exception() performs the
transition right away in all states. This is a massive handgun pointed right at
our metaphoric feet:

- just after calling exception() in startHook or updateHook, the component is
  essentially cleaned up. If one forgets to return and starts using some
  internal properties, it is probably going to crash
- when called in stopHook and cleanupHook, exception() would call stopHook
  and/or cleanupHook again, which is rather obviously not good: the second
  hook call will try to re-do things that have already been done, which
  could include freeing or using already freed resources, and it could
  lead to an infinite recursion (think I/O timeout on device lost)

This commit fixes the second bit. The hooks are now called after the initial
hook returned, that is instead of:

  stopHook
    -> exception
       -> stopHook
          cleanupHook
          exceptionHook

We now have

  stopHook
    -> exception
  cleanupHook
  exceptionHook

The first point is not (yet) handled.